+++ internal/gatewaysecret/client/client.go
