+++ internal/gatewaysecret/handler_test.go
@@ -14,7 +14,7 @@ import (
 	apimetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/kyma-project/lifecycle-manager/api/shared"
-	"github.com/kyma-project/lifecycle-manager/pkg/gatewaysecret"
+	"github.com/kyma-project/lifecycle-manager/internal/gatewaysecret"
 )
 
 func TestManageGatewaySecret_WhenGetGatewaySecretReturnsError_ReturnsError(t *testing.T) {
