+++ internal/gatewaysecret/mock_Client_test.go
