+++ unit-test-coverage.yaml
@@ -4,6 +4,7 @@ packages:
   internal/descriptor/cache: 92.3
   internal/descriptor/provider: 66.7
   internal/event: 100
+  internal/gatewaysecret: 100
   internal/manifest: 15.8
   internal/manifest/statecheck: 71
   internal/manifest/filemutex: 100
@@ -18,4 +19,3 @@ packages:
   internal/remote: 20.2
   internal/util/collections: 86
   pkg/templatelookup: 77.1
-  pkg/gatewaysecret: 100
